\name{BIMA-package}
\alias{BIMA-package}
\alias{BIMA}
\docType{package}
\title{
\packageTitle{BIMA}
}
\description{
\packageDescription{BIMA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BIMA}
\packageIndices{BIMA}
Using softh-thresholding Gaussian process prior to analyze high-dimensional image mediation effect
}
\author{BIMA
\packageAuthor{BIMA}

Maintainer: \packageMaintainer{BIMA}
}
\references{
See also Kang et. al. (2018)
}

\keyword{Mediation, Imaging, Bayes }
\seealso{
 Optional links to other man pages
}

