% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions_for_testing.R
\name{BIMA}
\alias{BIMA}
\title{Bayesian Image Mediation Analysis}
\usage{
BIMA(
  Y,
  X,
  M,
  C,
  dim = 2,
  grids = NULL,
  init_y = NULL,
  init_m = NULL,
  controls_y = NULL,
  controls_m = NULL,
  region_idx_list = NULL,
  kernel_setting = list(method = "Exponential_square", kernel_params = NULL, Phi_Q =
    NULL, Phi_D = NULL)
)
}
\arguments{
\item{Y}{The scalar outcome, n by 1}

\item{X}{The scalar exposure variable, n by 1}

\item{M}{The image predictor, n by p}

\item{C}{The q confounders, q by n}

\item{dim}{An integer to specify the dimension of the input image, where it is 1D 2D or 3D image.
This must be specified when kernel_setting = "Self-defined".}

\item{grids}{A matrix of size p by dim. For a 2D image,
\texttt{grids[j,]} represents the location (x-y coordinate) of pixel j. For 3D image, \texttt{grids[j,,]} represents
the location (x-y-z coordinates) of voxel j. This must be specified when kernel_setting = "Self-defined".}

\item{init_y}{A list of initial parameter settings for Scalar-on-image regression.
\itemize{
     \item theta_beta, A vector of length L, where L is the total number of basis functions. Default value is rep(1,L).
     \item a_sigma_beta Scalar, parameter a in IG(a,b) for the prior of sigma_Y and sigma_beta. Default value is 1
     \item b_sigma_beta Scalar, parameter b in IG(a,b) for the prior of sigma_Y and sigma_beta. Default value is 1
     \item sigma_Y Scalar,initial value for sigma_Y. Default value is 1.
     \item sigma_beta Scalar,initial value for sigma_beta. Default value is 1.
     \item cb Scalar, initial value for the intercept term. Default value is 0.
     \item zetay Vector of length q. Default value is rep(1,q).
     \item gamma Scalar, initial value for the coefficient of exposure variable X. Default value is 1.
}}

\item{init_m}{A list of intial parameter settings for Image-on-scalar regression.
\itemize{
    \item theta_alpha A vector of length L, where L is the total number of basis functions. Default value is rep(1,L).
    \item theta_eta A matrix of size L by n. Initial value for the inidividual effect eta_i. Default value is matrix(0, L,n)
    \item sigma_M Scalar, initial value for sigma_M. Default value is 1.
    \item sigma_alpha Scalar, initial value for sigma_alpha Default value is 1.
    \item sigma_eta Scalar, initial value for sigma_eta Default value is 0.1.
    \item theta_zetam A matrix of size L by m, initial value for the basis coefficients for the coefficient zetam for confounders C.
    Default value is  matrix(rep(0.1,L*m),L,m)).
}}

\item{controls_y}{A list of controls for running MALA algorithm for the scalar-on=image regression.
\itemize{
    \item lambda The thresholding parameter in STGP prior. Default value is 0.5
    \item n_mcmc Number of total MCMC iterations, default  = 1e5.
    \item stop_adjust The first number of iterations where step_size will be adjusted to achieve target acceptance rate,
    default = 0.8*1e5.
    \item start_joint The algorithm will update theta_beta alone for the first \texttt{start_joint} number of iterations.
     After that all parameters will be jointly updated. We only recommend setting start_joint>0 when the problem is very high-dimensional.
     default = 0.
    \item interval_step An integer, set the frequency of adjusting step size, default = 10
    \item interval_thin An integer, set the frequency of saving the MCMC for theta_beta, default = 1.
    \item step A numeric variable, initial step size. Default = 1e-2/n
}}

\item{controls_m}{A list of controls for running MALA algorithm for the image-on-scalar regression.
\itemize{
    \item lambda The thresholding parameter in STGP prior. Default value is 0.5
    \item n_mcmc Number of total MCMC iterations, default  = 2e4.
    \item stop_adjust The first number of iterations where step_size will be adjusted to achieve target acceptance rate,
    default = 0.8*2e4.
    \item start_joint start_joint The algorithm will update theta_alpha alone for the first \texttt{start_joint} number of iterations.
     After that all parameters will be jointly updated. We only recommend setting start_joint>0 when the problem is very high-dimensional.
     default = 0.
    \item interval_step An integer, set the frequency of adjusting step size, default = 10
    \item interval_eta An integer, set the frequency of updating the individual effect theta_eta.
    The accuracy of theta_eta does not effect the result too much. Larger interval_eta can potentially improve the computational speed.
    default = 1000.
    \item thinning An integer, set the frequency of saving theta_alpha, default = 1
    \item step A numeric variable, initial step size. Default = 1e-2/n
}}

\item{region_idx_list}{A list of length num_region. Users are allowed to pre-specify
the region parcellation by putting indices for each region as each component in this list.
The dafault value is list(region1 = 1:p) with only 1 region.}

\item{kernel_setting}{A list to specify the kernel setting by the following components
\itemize{
    \item method A character string to specify the type of kernel used. Currently implemented for
    Exponential_square for the modified exponential square kernel and Matern for matern kernel.
    Users can also set \texttt{method = "Self-defined"}, and directly use the basis functions and eigenvalues
    by putting them in kernel_setting$Phi_Q and kernel_setting$Phi_D respectively.

    \item kernel_params A list object.
    If \texttt{method = "Exponential_square"}, the default value is \texttt{kernel_params = list(a = 0.01,b=10,poly_degree = 10)},
    where a and b are the parameters in modified exponential square kernel \sqn{K(s,t) = \exp(-a(s^2+t^2) - b(s-t)^2)}.
    poly_degree is an integer number specifying the highest degree of Hermite polynomials.
    If \texttt{method = "Matern"}, the default value is \texttt{kernel_params = list(L_vec = 0.5*unlist(lapply(region_idx,length)), scale = 2, nu = 1/5)}.
    L_vec is a vector of integers specifying the number of basis functions used in each region.
    See wikipedia for the definition of matern kernel.
    Scale represents \sqn{\rho} in this definition, and nu represents \sqn{\nu}.

    \item Phi_Q A list object with length num_region.
    Each component represents one basis function, a matrix of dimension p_r by L_r for the r-th region.
    Needs to be specified if method = "Self-defined".
    \item Phi_D A list object with length num_region.
    Each component represents one set of eigenvalues for one region, a vector of length L_r for the r-th region.
    Needs to be specified if method = "Self-defined".
}}
}
\value{
A list of MCMC samples. The default will output the last 20\% as the samples after burn-in.
\itemize{
   \item beta_sample
   \item alpha_sample
   \item NIE_sample MCMC sample for the natural indirect effect, product of alpha*beta
   \item gamma_sample MCMC sample for the natural direct effect
}
}
\description{
BIMA function is an illustrative example for working with small data sets.
For large scale imaging mediation analysis, please run Y_regression_region_block_fast and
M_regression_region_block separately for the scalar-on-image and image-on-scalar regressions.
These 2 main functions need to be tuned separately to achive best performance.
}
\examples{
if(FALSE){
    data(one_region)
    # see also ./docs/small_example.R
}
}
