% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIMA_package.R
\docType{package}
\name{BIMA_pkg}
\alias{BIMA_pkg}
\alias{BIMA-package}
\title{BIMA package}
\description{
Using softh-thresholding Gaussian process prior to analyze high-dimensional image mediation effect.
}
