% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{datsim}
\alias{datsim}
\title{datsim: A simulated data set for image mediation analysis}
\format{
A list object that contains both the data and the true parameters. n = 300, p=400
\describe{
  \item{Y}{scalar outcome, n by 1}
  \item{M}{Image mediator, p by n}
  \item{X}{Exposure variable, n by 1}
  \item{C}{q Confounder variables, q by n}
  \item{....}{All other components are true parameters used to generate the data set. }
}
}
\description{
This data set is generated by STGP_generate_theta_block function using the same cammond as in ./docs/BIMA_example.Rmd
This data sets contains 4 regions for 2D image.
}
