% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions_for_testing.R
\name{GP.simulate.curve.fast.new}
\alias{GP.simulate.curve.fast.new}
\title{An alternative version to GP.simulate.curve.fast in BayesGPfit}
\usage{
GP.simulate.curve.fast.new(
  x,
  poly_degree,
  a,
  b,
  center = NULL,
  scale = NULL,
  max_range = 6
)
}
\arguments{
\item{x}{A p by d matrix, denoting the grids of a d-dim image with p pixels.
This is usually produced by BayesGPfit::GP.generate.grids}

\item{poly_degree}{Integer, polynomial degrees.}

\item{a}{Parameter a in GP kernel \eqn{K(s,t) = exp(-a^2*(s^2+t^2)-b*(s-t)^2 }}

\item{b}{Parameter b in GP kernel \eqn{K(s,t) = exp(-a^2*(s^2+t^2)-b*(s-t)^2 }}

\item{center}{Center of the grid}

\item{scale}{Parameter for scaling the distance between grid points}

\item{max_range}{A scaling parameter, default at 6.}
}
\value{
a list
\itemize{
\item
}
}
\description{
This function is used to generate basis functions and eigenvalues
}
