% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{one_region}
\alias{one_region}
\title{one_region: A simulated data with 1 region for image mediation analysis}
\format{
A list object that contains both the data and the true parameters. n = 300, p=400
\describe{
  \item{Y}{scalar outcome, n by 1}
  \item{M}{Image mediator, p by n}
  \item{X}{Exposure variable, n by 1}
  \item{C}{q Confounder variables, q by n}
  \item{beta_test_ST}{True image used to generate beta}
  \item{alpha_test_ST}{True image used to generate alpha}
  \item{....}{All other components are true parameters used to generate the data set. }
}
}
\description{
This data set is generated by STGP_generate_theta_block function using the same cammond as in ./docs/small_example.Rmd
This data sets contains 1 regions for 2D image.
}
