% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Y_regression_region_block_fast}
\alias{Y_regression_region_block_fast}
\title{Scalar on image regression}
\usage{
Y_regression_region_block_fast(
  Y,
  M,
  X,
  C,
  L_all,
  num_region,
  region_idx,
  n_mcmc,
  K,
  stop_burnin,
  start_joint,
  lambda,
  target_accept_vec,
  a,
  b,
  init,
  step,
  interval_step,
  interval_thin,
  display_progress = TRUE
)
}
\arguments{
\item{Y}{The scalar outcome, n by 1}

\item{M}{The image predictor, n by p}

\item{X}{The scalar exposure variable, n by 1}

\item{C}{The q confounders, n by q}

\item{L_all}{A vector of length num_region, each element is an integer to indicate the number of basis in each region}

\item{num_region}{An integer, the total number of regions}

\item{region_idx}{A list object of length num_region, each element is a vector of
the indices of each voxel in that region. Note that this index starts from 0.}

\item{n_mcmc}{An integer to indicate the total number of MCMC iterations}

\item{K}{A list object of length num_region, the r-th element is a p_r by L_r matrix for the basis function}

\item{stop_burnin}{An integer to indicate from which iteration to stop burnin period.
Note that during burinin, the step size in MALA is adjusted every interval_step iterations.}

\item{start_joint}{An integer to indicate from which iteration to start join update all parameters.
Note that this function allows to update beta using MALA alone at the beginning.
Users may wait for a few iterations for beta to stabilize and then jointly update all other parameters (using Gibbs Sampling)}

\item{lambda}{A numeric variable to indicate the thresholding parameter lambda in STGP prior}

\item{target_accept_vec}{A vector of length num_region. Each element is a numeric variable in (0,1).
This allows the user to define different target acceptance rate for each region in the MALA algorithm,
and the step size will be adjusted to meet the target acceptance rate.}

\item{a}{A numeric variable for the Inverse-Gamma(a,b), priors for \eqn{\sigma^2_Y,\sigma^2_\beta}}

\item{b}{A numeric variable for the Inverse-Gamma(a,b), priors for \eqn{\sigma^2_Y,\sigma^2_\beta}}

\item{init}{A list object that contains the following element
\itemize{
  \item theta_beta A vector of length L. Initial value for theta_beta
  \item D A vector of length L. Eigenvalues for all regions in the basis
  \item gamma A numeric scalar, initial value for gamma
  \item cb A numeric scalar, initial value for the intercept term
  \item zetay A vector of length q, intial value for zetay
  \item sigma_Y A numeric scalar, intial value for sigma_Y
  \item sigma_beta A numeric scalar, initial value for sigma_beta
}}

\item{step}{A numeric vector of length num_region, the initial step size for each region}

\item{interval_step}{An integer to denote how often to update the step size}

\item{interval_thin}{An integer to denote how often to save the thinned MCMC sample for theta_beta}

\item{display_progress}{True for displaying progress bar}
}
\value{
A List object with the following component
\itemize{
\item theta_beta_mcmc_thin
\item logll_mcmc_Y
\item track_step
\item accept_block
\item emp_accept
\item gs A list object with the following component
  \itemize{
    \item zetay_mcmc
    \item cy_mcmc
    \item sigma_beta2_mcmc
    \item sigma_Y2_mcmc
  }
\item time_seg
\item timer
}
}
\description{
A basis decomposition is used. The main coefficient beta follows STGP prior.
Kernel matrices need to be prespeficified
}
